requirements.txt
langchain
openai
python-dotenv
langchain-openai
huggingface_hub
streamlit